Move formatRRuleText to calendar package
