+++ packages/calendar/package.json
@@ -19,6 +19,7 @@
     "generate-disclaimer": "echo \"Nothing to generate\""
   },
   "dependencies": {
+    "i18next": "^23.7.14",
     "joi": "^17.11.0",
     "lodash": "^4.17.20",
     "luxon": "^3.3.0",
