+++ packages/calendar/src/lib/utils/helpers.ts
